reputation: move unit tests
UTs are moved inside "test/" directory,
and reworked to improve readability and reduce lines of code.
